Some warning fixes.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@30549 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/c64/cart/rrnetmk3.c
@@ -145,7 +145,7 @@ void rrnetmk3_reset(void)
 #ifdef HAVE_TFE
     tfe_clockport_changed();
 #endif
-    cart_config_changed_slotmain(CMODE_RAM, rrnetmk3_biossel ? CMODE_RAM : CMODE_8KGAME, CMODE_READ);
+    cart_config_changed_slotmain(CMODE_RAM, (BYTE)(rrnetmk3_biossel ? CMODE_RAM : CMODE_8KGAME), CMODE_READ);
 }
 
 static int set_rrnetmk3_flashjumper(int val, void *param)
@@ -184,7 +184,7 @@ void rrnetmk3_config_init(void)
 {
     LOG(("RRNETMK3 rrnetmk3_config_init"));
     rrnetmk3_biossel = rrnetmk3_hw_flashjumper; /* disable bios at reset when flash jumper is set */
-    cart_config_changed_slotmain(CMODE_RAM, rrnetmk3_biossel ? CMODE_RAM : CMODE_8KGAME, CMODE_READ);
+    cart_config_changed_slotmain(CMODE_RAM, (BYTE)(rrnetmk3_biossel ? CMODE_RAM : CMODE_8KGAME), CMODE_READ);
 }
 
 static void rrnetmk3_io1_store(WORD addr, BYTE value)
